Start on LogHandler class and implementation in tests
