+++ src/test/java/org/broadinstitute/consent/http/resources/DataRequestVoteResourceTest.java
@@ -17,6 +17,7 @@ import org.broadinstitute.consent.http.service.ElectionService;
 import org.broadinstitute.consent.http.service.EmailNotifierService;
 import org.broadinstitute.consent.http.service.UserService;
 import org.broadinstitute.consent.http.service.VoteService;
+import org.broadinstitute.consent.http.util.LogHandler;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -32,6 +33,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
@@ -102,7 +105,7 @@ public class DataRequestVoteResourceTest {
         return dar;
     }
 
-    private void testCreateDataOwnerElection(Vote vote, User user, boolean returnEmptyList) throws Exception {
+    private void enableCreateDataOwnerElection(Vote vote, User user, boolean returnEmptyList) throws Exception {
         vote.setElectionId(1);
         if (returnEmptyList) {
             when(voteService.describeVoteByTypeAndElectionId(any(), any())).thenReturn(Collections.emptyList());
@@ -229,16 +232,22 @@ public class DataRequestVoteResourceTest {
         User user = createMockUser(role, 1);
         when(userService.findUserByEmail(any())).thenReturn(user);
 
-        Vote vote = createMockVote(VoteType.DATA_OWNER.getValue(), 1, true);
+        Vote vote = createMockVote("", 1, true);
         when(voteService.findVoteById(any())).thenReturn(vote);
+
+        Logger logger = Logger.getLogger(DataRequestVoteResource.class.getName());
+        LogHandler handler = new LogHandler(Level.ALL);
+        logger.setUseParentHandlers(false);
+        logger.addHandler(handler);
+        logger.setLevel(Level.ALL);
 
         when(voteService.updateVoteById(any(), any())).thenReturn(vote);
         when(electionService.validateCollectDAREmailCondition(any())).thenReturn(true);
         doThrow(new IOException()).when(emailNotifierService).sendCollectMessage(any());
-        //TODO: Check logger to verify "severe" status has been posted (issue DUOS-1515)
 
         initResource();
         Response response = resource.createDataRequestVote(authUser, uriInfo, "", 1, "");
+        assertEquals(Level.SEVERE, handler.checkLevel());
         assertEquals(200, response.getStatus());
     }
 
@@ -312,7 +321,7 @@ public class DataRequestVoteResourceTest {
         DataAccessRequest dar = createMockDAR();
         when(dataAccessRequestService.findByReferenceId(any())).thenReturn(dar);
 
-        testCreateDataOwnerElection(vote, user, true);
+        enableCreateDataOwnerElection(vote, user, true);
 
         initResource();
         String json = "{\"electionId\":1,\"vote\":true}";
@@ -334,7 +343,7 @@ public class DataRequestVoteResourceTest {
         DataAccessRequest dar = createMockDAR();
         when(dataAccessRequestService.findByReferenceId(any())).thenReturn(dar);
 
-        testCreateDataOwnerElection(vote, user, false);
+        enableCreateDataOwnerElection(vote, user, false);
 
         initResource();
         String json = "{\"electionId\":1,\"vote\":true}";
